imports updated for moved Pattern, Label classes new .gitignore with rule to ignore .pyc files new __init__.py files to make packages accessible
