+++ src/core/epanet/hydraulics/node.py
@@ -1,6 +1,6 @@
 ﻿from enum import Enum
 
-from core.epanet.time_patterns.patterns import Pattern
+from core.epanet.patterns import Pattern
 from core.epanet.curves import Curve
